% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenaire-package.R
\docType{package}
\name{ropenaire-package}
\alias{ropenaire-package}
\alias{ropenaire}
\title{\strong{Client for the OpenAIRE API}}
\description{
\strong{Client for the OpenAIRE API}
}
\author{
Najko Jahn \email{najko.jahn@uni-bielefeld.de}

Scott Chamberlain \email{myrmecocystus@gmail.com}
}
