% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roa_datasets.R
\name{roa_datasets}
\alias{roa_datasets}
\title{Search datasets}
\usage{
roa_datasets(fp7 = NULL, limit = 1000, id = NULL, title = NULL,
  author = NULL, from_date = NULL, to_date = NULL, ...)
}
\arguments{
\item{fp7}{Search for dataset associated to a FP7 project with the given grant number}

\item{limit}{limit number of records}

\item{id}{Gets the dataset with the given openaire identifier, if any.}

\item{title}{Publication title}

\item{author}{Search for dataset by authors}

\item{from_date}{Gets the dataset whose date of acceptance is greater than or equal the given date.
Allowed values: date formatted as YYYY-MM-DD.}

\item{to_date}{Gets the dataset whose date of acceptance is less than or equal the given date.
Allowed values: date formatted as YYYY-MM-DD.}

\item{...}{other API parameters \url{http://api.openaire.eu/}}
}
\description{
Search datasets
}
\details{
OpenAIRE is collecting the metadata of research datasets
that are related to an open access publication in OpenAIRE or linked to a funding stream
}
\examples{
\dontrun{
# Search for FP7 grant-supported datasets
roa_datasets(fp7 = "247153")

roa_datasets(title = "methane")
}
}

