% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roa_projects.R
\name{roa_projects}
\alias{roa_projects}
\title{Get EC projects}
\usage{
roa_projects(org = NULL, limit = 1000, ...)
}
\arguments{
\item{org}{Search for projects by participant institutions (acronym)}

\item{limit}{number of projects to be returned}

\item{...}{other API parameters \url{http://api.openaire.eu/}}
}
\description{
Get EC projects
}
\examples{
roa_projects(org = "UGOE")
}

