% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roa_projects.R
\name{roa_projects}
\alias{roa_projects}
\title{Get EC projects}
\usage{
roa_projects(grant_id = NULL, publication_id = NULL, title = NULL,
  acronym = NULL, call_id = NULL, start_year = NULL, end_year = NULL,
  country = NULL, org = NULL, size = 1000, sort_by = NULL,
  sort_order = NULL, ...)
}
\arguments{
\item{grant_id}{Gets the project with the given grant identifier,
if any}

\item{publication_id}{publication ID}

\item{title}{Name of the project}

\item{acronym}{Gets the project with the given acronym, if any.}

\item{call_id}{Search for projects by call identifier}

\item{start_year}{Gets the projects that started in the given year.
Format: YYYY}

\item{end_year}{Gets the projects that ended in the given year.
Format: YYYY}

\item{country}{Search for projects by participant countries. Format:
2-letter country code}

\item{org}{Search for projects by participant institutions (acronym)}

\item{size}{size number of records}

\item{sort_by}{sort field, one of projectstartdate, projectstartyear,
projectenddate, projectendyear, projectduration}

\item{sort_order}{sort order. one of ascending or descending. optional}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Get EC projects
}
\note{
\code{format} is hard coded to \code{xml} for now
}
\examples{
\dontrun{
roa_projects(org = "UGOE", size = 3)
roa_projects(call_id = "FP7-PEOPLE-2010-IRSES")
roa_projects(title = "open", size = 3)

# curl options
x <- roa_projects(org = "UGOE", size = 10, verbose = TRUE)
}
}
\references{
OpenAIRE API docs \url{http://api.openaire.eu/}
}
