% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roa_pubs.R
\name{roa_pubs}
\alias{roa_pubs}
\title{Search FP7 publications}
\usage{
roa_pubs(fp7 = NULL, limit = 1000, ...)
}
\arguments{
\item{fp7}{Search for publications associated to a FP7 project with the given grant number}

\item{limit}{limit number of records}

\item{...}{other API parameters \url{http://api.openaire.eu/}}
}
\description{
Search FP7 publications
}

