% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roa_pubs.R
\name{roa_pubs}
\alias{roa_pubs}
\title{Search FP7 publications}
\usage{
roa_pubs(fp7 = NULL, publication_id = NULL, doi = NULL,
  provider_id = NULL, project_id = NULL, openaire_project_id = NULL,
  has_project = NULL, oa = NULL, title = NULL, author = NULL,
  from_date = NULL, to_date = NULL, model = NULL,
  fp7_scientific_area = NULL, has_ec_funding = NULL,
  has_wt_funding = NULL, funder = NULL, funding_stream = NULL,
  keywords = NULL, size = 1000, page = 1, sort_by = NULL,
  sort_order = NULL, format = "tsv", raw = FALSE, ...)
}
\arguments{
\item{fp7}{(character) Search for publications associated to a FP7 project
with the given grant number}

\item{publication_id}{(character) Gets the publication with the given
openaire identifier, if any.}

\item{doi}{(character) Gets the publications with the given DOIs}

\item{provider_id}{(character) Search for datasets by openaire data provider
identifier. Alternatevely, it is possible to repeat the parameter for each
provider id. In both cases, provider identifiers will form a query with OR
semantics.}

\item{project_id}{(character) The given grant identifier of the project.}

\item{openaire_project_id}{(character) Search for publications by openaire project
identifier. Alternatevely, it is possible to repeat the parameter for each
provider id. In both cases, provider identifiers will form a query with OR
semantics.}

\item{has_project}{(logical) If hasProject is true gets the datasets that
have a link to a project. If hasProject is false gets the datasets with no
links to projects.}

\item{oa}{(logical) Open access or not.}

\item{title}{(character) Publication title}

\item{author}{(character) Search for publications by authors}

\item{from_date}{(character) Gets the publications whose date of acceptance
is greater than or equal the given date. Allowed values: date formatted as
YYYY-MM-DD.}

\item{to_date}{(character) Gets the publications whose date of acceptance
is less than or equal the given date. Allowed values: date formatted as
YYYY-MM-DD.}

\item{model}{(character) The data model of the response. Options: 
"openaire" or "sygma". Model "sygma" is a simplified version of the 
"openaire" model. For sygma, only the xml format is available. See 
the relative XML schema 
(<https://www.openaire.eu/schema/sygma/oaf_sygma_v2.1.xsd>). 
Default: "openaire". Unfortunately, passing "model=openaire" leads to
no results AFAICT, so this parameter is left as `NULL` for now.
If you pass `model=sygma` we force `format=xml` (see above)}

\item{fp7_scientific_area}{(character) Search for FP7 entities by 
scientific area}

\item{has_ec_funding}{(logical) If `TRUE` gets the entities funded by the 
EC. If `FALSE` gets the entities related to projects not funded by the EC.}

\item{has_wt_funding}{(logical) If `TRUE` gets the entities funded by 
Wellcome Trust. The results are the same as those obtained with 
`funder="wt"`. If `FALSE` gets the entities related to projects not 
funded by Wellcome Trust.}

\item{funder}{(character) Search for entities by funder. One of WT, EC, 
ARC, ANDS, NSF, FCT, NHMRC}

\item{funding_stream}{(character) xx}

\item{keywords}{(character) vector of keywords}

\item{size}{(integer) number of records}

\item{page}{(integer) page number of the search results}

\item{sort_by}{(character) sort field, one of dateofcollection,
resultstoragedate, resultstoragedate, resultembargoenddate,
resultembargoendyear, resultdateofacceptance, resultacceptanceyear}

\item{sort_order}{(character) sort order. one of ascending or descending. 
optional}

\item{format}{(character) format to return, one of json, xml, csv 
or tsv (default)}

\item{raw}{(logical) return raw text or not. Default: \code{FALSE}}

\item{...}{curl options passed on to [crul::HttpClient]}
}
\description{
Search FP7 publications
}
\examples{
\dontrun{
# Search for FP7 grant-supported publications
my_pubs <- roa_pubs(fp7 = "283595")
# Show how many of them are Open Access
table(my_pubs$Access)

# Search by doi
roa_pubs(doi = "10.1051/0004-6361/201220935")

# has_project boolean
roa_pubs(has_project = FALSE, size = 10)
roa_pubs(has_project = TRUE, size = 10)

# curl options
x <- roa_pubs(doi = "10.1051/0004-6361/201220935", verbose = TRUE)
}
}
\references{
OpenAIRE API docs \url{http://api.openaire.eu/}
}
