% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roa_pubs.R
\name{roa_pubs}
\alias{roa_pubs}
\title{Search FP7 publications}
\usage{
roa_pubs(fp7 = NULL, publication_id = NULL, dataset_id = NULL,
  doi = NULL, provider_id = NULL, project_id = NULL, has_project = NULL,
  oa = NULL, title = NULL, author = NULL, from_date = NULL,
  to_date = NULL, size = 1000, sort_by = NULL, sort_order = NULL,
  format = "tsv", raw = FALSE, ...)
}
\arguments{
\item{fp7}{Search for publications associated to a FP7 project with the
given grant number}

\item{publication_id}{Gets the publication with the given openaire
identifier, if any.}

\item{dataset_id}{Gets the dataset with the given openaire identifier,
if any}

\item{doi}{Gets the publications with the given DOIs}

\item{provider_id}{Search for datasets by openaire data provider identifier.
Alternatevely, it is possible to repeat the parameter for each provider id.
In both cases, provider identifiers will form a query with OR semantics.}

\item{project_id}{Search for publications by openaire project identifier.
Alternatevely, it is possible to repeat the parameter for each provider id.
In both cases, provider identifiers will form a query with OR semantics.}

\item{has_project}{If hasProject is true gets the datasets that have a link
to a project. If hasProject is false gets the datasets with no links to
projects.}

\item{oa}{Open access or not.}

\item{title}{Publication title}

\item{author}{Search for publications by authors}

\item{from_date}{Gets the publications whose date of acceptance is greater
than or equal the given date. Allowed values: date formatted as YYYY-MM-DD.}

\item{to_date}{Gets the publications whose date of acceptance is less than
or equal the given date. Allowed values: date formatted as YYYY-MM-DD.}

\item{size}{size number of records}

\item{sort_by}{sort field, one of dateofcollection, resultstoragedate,
resultembargoenddate, resultembargoendyear, resultdateofacceptance,
resultacceptanceyear}

\item{sort_order}{sort order. one of ascending or descending. optional}

\item{format}{format to return, one of json, xml, csv or tsv (default)}

\item{raw}{(logical) return raw text or not. Default: \code{FALSE}}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Search FP7 publications
}
\examples{
\dontrun{
# Search for FP7 grant-supported publications
my_pubs <- roa_pubs(fp7 = "283595")
# Show how many of them are Open Access
table(my_pubs$Access)

# publication id
# roa_pubs(publication_id = )

# dataset id
# roa_pubs(dataset_id = )

# Search by doi
roa_pubs(doi = "10.1051/0004-6361/201220935")

# provider id
# roa_pubs(provider_id = )

# project id
# roa_pubs(project_id = )

# has_project boolean
roa_pubs(has_project = FALSE, size = 10)
roa_pubs(has_project = TRUE, size = 10)

# publication id
# roa_pubs(publication_id = )


# curl options
x <- roa_pubs(doi = "10.1051/0004-6361/201220935", verbose = TRUE)
}
}
\references{
OpenAIRE API docs \url{http://api.openaire.eu/}
}
